Checking if the user is the owner of a workspace before allowing it to do stuff

+++ controllers/workspaceController.ts
@@ -1,6 +1,7 @@
 import express from "express";
 import logger from "../lib/logger";
 import { WorkspaceService } from "../services/workspaceService";
+import { RequestWithUser } from "../dataModels/AuthenticationModels";
 const router = express.Router();
 
 /**
@@ -55,15 +56,22 @@ router.get("/", async (req, res) => {
  *                                                               
  */
 
-router.get("/:id", async (req, res) => {
+router.get("/:id", async (req: RequestWithUser, res) => {
     try{
         const { id } = req.params;
+        const { user } = req;
+        if(!(await WorkspaceService.isOwner(user.userId, id))) 
+            throw new Error("Forbidden: User has no permission to access this workspace");
+
         const workspace = await WorkspaceService.getById(id);
         if(workspace) return res.status(200).json(workspace);
 
         return res.status(404).send();
     }catch(error: any){
         logger.error(error.message);
+        if(error.message === "Forbidden: User has no permission to access this workspace") 
+            return res.status(403).json({error: error.message});
+        
         return res.status(500).json({error: error.message});
     }
 });
@@ -94,19 +102,26 @@ router.get("/:id", async (req, res) => {
  *                                
  */
 
-router.post("/", async (req, res) => {
+router.post("/", async (req: RequestWithUser, res) => {
     const { body } = req;
+    const { user } = req;
     try {
         if (
-            "userId" in body && typeof body.userId === "string" &&
             "name" in body && typeof body.name === "string"
         ) {
+            if(user.userId === undefined) 
+                throw new Error("Forbidden: User has no permission to create a workspace");
+
+            body.userId = user.userId;
+
             const workspace = await WorkspaceService.create(body);
             return res.status(200).json(workspace);
         }
         return res.status(400).json({ message: "Invalid request" });
     } catch (error: any) {
         logger.error(error.message);
+        if(error.message === "Forbidden: User has no permission to create a workspace")
+            return res.status(403).json({error: error.message});
         return res.status(500).json({error: error.message});
     }
 });
@@ -144,16 +159,22 @@ router.post("/", async (req, res) => {
  *                                
  */
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", async (req: RequestWithUser, res) => {
     const { id } = req.params;
     const { body } = req;
+    const { user } = req;
     try {
+        if(!(await WorkspaceService.isOwner(user.userId, id))) 
+            throw new Error("Forbidden: User has no permission over this workspace");
+
         const workspace = await WorkspaceService.update(id, body);
         if(workspace) return res.status(200).json(workspace);
 
         return res.status(404).json();
     }catch(error: any){
         logger.error(error.message);
+        if(error.message === "Forbidden: User has no permission over this workspace")
+            return res.status(403).json({error: error.message});
         return res.status(500).json({error: error.message});
     }
 });
@@ -185,15 +206,21 @@ router.put("/:id", async (req, res) => {
  *                                
  */
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", async (req: RequestWithUser, res) => {
     const { id } = req.params;
+    const { user } = req;
     try{
+        if(!(await WorkspaceService.isOwner(user.userId, id))) 
+            throw new Error("Forbidden: User has no permission over this workspace");
+
         const workspace = await WorkspaceService.delete(id);
         if(workspace) return res.status(200).json(workspace);
         
         return res.status(404).json();
     }catch(error: any){
         logger.error(error.message);
+        if(error.message === "Forbidden: User has no permission over this workspace")
+            return res.status(403).json({error: error.message});
         return res.status(500).json({error: error.message});
     }
 });
